factor out bookserror enum
